Rework _parse_route_data to allow multiple routing tables.
I think it is a bird2 feature, but not sure. I tried parsing a system
with ipv6 enabled, which burned down because a line with 'Table:
master6' suddenly appeared unexpedted. I guess this is a bird2 feature.
I have reworked the _parse_route_data a bit more pythonic and to cope
with this.

I have introdcuced a dependancy, more_itertools. If that is undesired
i'll go back to the drawingboard.